\name{plot_posterior}
\alias{plot_posterior}
\title{Plot fitted posterior summaries}
\usage{
plot_posterior(object)
}
\arguments{
\item{object}{A \code{vcmoe} object.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots mean posterior probabilities for each component over the coefficient grid.
}
